Rename & remove testing request data
